+++ packages/pds/package.json
@@ -60,12 +60,7 @@
     "@opentelemetry/api": "^1.9.1",
     "@opentelemetry/api-logs": "^0.219.0",
     "@opentelemetry/instrumentation-aws-sdk": "^0.74.0",
-    "@opentelemetry/instrumentation-express": "^0.67.0",
-    "@opentelemetry/instrumentation-http": "^0.219.0",
     "@opentelemetry/instrumentation-ioredis": "^0.67.0",
-    "@opentelemetry/instrumentation-pino": "^0.65.0",
-    "@opentelemetry/instrumentation-runtime-node": "^0.32.0",
-    "@opentelemetry/instrumentation-undici": "^0.29.0",
     "better-sqlite3": "^12.0.0",
     "compression": "^1.7.4",
     "cors": "^2.8.5",
